feat(pre-commit): setup pre-commit and use black as the default formatter

+++ nnsmith/util.py
@@ -7,15 +7,17 @@ import numpy as np
 
 def mkdir(dir, yes=False):
     if os.path.exists(dir):
-        decision = ''
+        decision = ""
         if yes:
-            decision = 'y'
-        while decision.lower() not in ['y', 'n']:
+            decision = "y"
+        while decision.lower() not in ["y", "n"]:
             decision = input(
-                'Report folder already exists. Press [Y/N] to continue or exit...')
-        if decision.lower() == 'n':
+                "Report folder already exists. Press [Y/N] to continue or exit..."
+            )
+        if decision.lower() == "n":
             raise RuntimeError(
-                f'{dir} already exist... We want an empty folder to report...')
+                f"{dir} already exist... We want an empty folder to report..."
+            )
         else:
             shutil.rmtree(dir)
 
@@ -27,8 +29,9 @@ def gen_one_input(inp_spec, l, r, seed=None):
         np.random.seed(seed)  # TODO: use standalone random generator
     inp = {}
     for name, shape in inp_spec.items():
-        inp[name] = np.random.uniform(
-            low=l, high=r, size=shape.shape).astype(shape.dtype)
+        inp[name] = np.random.uniform(low=l, high=r, size=shape.shape).astype(
+            shape.dtype
+        )
     return inp
